+++ packages/@tinacms/graphql/src/mdx/stringify.ts
@@ -439,11 +439,19 @@ ${out}
       }
 
     case 'text':
-      return {
-        type: 'text',
-        // @ts-ignore
-        value: node.text || '',
+      let returnNode: Content = { type: 'text', value: node.text || '' }
+
+      if (node?.bold) {
+        returnNode = { type: 'strong', children: [returnNode] }
       }
+      if (node?.italic) {
+        returnNode = {
+          type: 'emphasis',
+          children: [returnNode],
+        }
+      }
+      return returnNode
+
     default:
       console.log(`Unrecognized field type: ${node.type}`)
       break